+++ app/mb/views/main.py
@@ -101,6 +101,7 @@ from mb.models import (
     SourceHabitat,
     MasterHabitat)
 from mb.models.location_models import MasterLocation, LocationRelation, SourceLocation
+from mb.utils.master_location_tools import MasterLocationView, get_master_habitats, string_contains, remove_none_values, filter
 
 
 # Error Pages
@@ -150,95 +151,8 @@ def index_proximate_analysis(request):
     return render(request, 'mb/index_proximate_analysis.html', context={'num_PA_item':num_PA_item},)
 
 def index_master_location_list(request):
-
-    def get_master_habitats(ml : MasterLocation):
-        """ Get MasterHabitats by MasterLocation """ 
-        mr = ml.reference
-
-        try:
-            master_habitats = MasterHabitat.objects.filter(reference=mr)
-        except:
-            return "nan"
-        
-        habitats = ""
-
-        for master_habitat in master_habitats:
-            habitats = habitats + f"{master_habitat.name} "
-
-        return str(habitats)
-
-    def string_contains(str1, str2):
-        """ If string contains substring. Converts string to lowercase. """
-        print("str1: " + str(str1))
-        print("str2: " + str(str2))
-
-        if str1.lower() in str2.lower():
-            return True
-        return False
-    
-    def remove_none_values(list):
-        """ Delete None-values from list. """
-        new_list = []
-        for item in list:
-            if item == None:
-                continue
-            new_list.append(item)
-        return new_list
-
-    def filter(objects, params):
-        """ This operates filter from web site. """
-        print("params: " + str(params))
-        
-        try:
-            master_location = str(params["master_location"])
-            for i in range(len(objects)):
-                if master_location == "":
-                    break
-                if string_contains(master_location, objects[i].name) == False:
-                    objects[i] = None
-        except Exception as e:
-            print("Virhe: " + str(e))
-        
-        objects = remove_none_values(objects)
-        try:
-            reference = str(params["reference"])
-            for i in range(len(objects)):
-                if reference == "":
-                    break
-                if string_contains(reference, objects[i].reference) == False:
-                    objects[i] = None
-        except Exception as e:
-            print("Virhe: " + str(e))
-        
-        
-        objects = remove_none_values(objects)
-        try:
-            master_habitat = params["master_habitat"]
-            for i in range(len(objects)):
-                if master_habitat == "":
-                    break
-                if string_contains(master_habitat, objects[i].master_habitat) == False:
-                    objects[i] = None
-        except Exception as e:
-            print("Virhe: " + str(e))
-		
-
-        objects = remove_none_values(objects)
-        for object in objects:
-            print("name: " + str(object.name))
-            continue
-        
-        return objects
-
-    params_set = False
-
     params = request.GET.dict()
 
-    class MasterLocationView(models.Model):
-        name = models.CharField(max_length=500) 
-        reference = models.CharField(max_length=500) 
-        master_habitat = models.TextField()
-    
     master_locations = MasterLocation.objects.filter()
 
     mls_with_habitat = []
