+++ app/tests/imports/test_master_location_filter.py
@@ -1,0 +1,5 @@
+from django.test import TestCase
+from mb.utils import *
+
+class MasterLocationFilterTest(TestCase):
+    pass
